CMUpdater: Add referrer in request header
XDA updated so that now a referrer needs to be passed in in order to get
a download. The referrer is the download URL without the "&task=get" at
the end of it.

Note that this now expects the updater API to only use XDA URLs, and it
expects the download URL without "&task=get" in it.